Create app1.py
The users file contains the current users and their reservations. An example of the format is as follows:
YesNo: Jan2_21 C0C1C2, April10_21 B0B1B2
WhatUp: Jan2_21 , April10_21 A0A1A2

The dates are the concerts, and the letters+numbers that follow are the seat reservations (by row letter and column number), if any. 

usersDict is a Python dictionary used throughout the program. This snippet of code stores the file's contents into usersDict at the beginning, so that it can be used within the program each time it is run. usersDict contains a dictionary within for each of the user key values, which in turn hold the seat reservations in each of the concert key values (making it a nested dictionary). This allows for easy look-up and storage of each user's concerts.