Create app2.py
Cancels all the reservations a user made for a particular concert. This used multiple files, one row of seats per file. Thus, I found no need to include more than A and B, given it is the same code. If the seat that the program is currently reading is taken by the user, then it is changed to "None," as in, empty.